Re-add proper Error type for whole client
