+++ light-spike/src/components/state.rs
@@ -1,17 +1,5 @@
 use crate::prelude::*;
 
-use thiserror::Error;
-
-#[derive(Debug, Clone, Error)]
-pub enum ErrorKind {
-    #[error("I/O error")]
-    Io(#[from] IoError),
-    #[error("no initial trusted state")]
-    NoInitialTrustedState,
-}
-
-pub type Error = anomaly::Error<ErrorKind>;
-
 #[derive(Debug)]
 pub struct State {
     pub trusted_store_reader: StoreReader<Trusted>,
