+++ light-spike/src/prelude.rs
@@ -8,9 +8,11 @@ pub use crate::components::io::*;
 pub use crate::components::scheduler::*;
 pub use crate::components::state::*;
 pub use crate::components::verifier::*;
+pub use crate::errors::*;
 pub use crate::operations::*;
 pub use crate::predicates::errors::*;
 pub use crate::predicates::VerificationPredicates;
 pub use crate::store::*;
 pub use crate::types::*;
+pub use crate::Never;
 pub use crate::{bail, ensure, postcondition, precondition};
